Add provided assignee option for import
Adds a "[Provided assignee]" option in CLI import to use the existing
assignee value from the imported issue.

Fixes an issue where a provided assignee value would result in
assigning the current user.

Moves the "[Unassigned]" and new "[Provided assignee]" options
at the start of the assignee options list.

Updates the README to detail the format expected for a value in an
Assignee column.