Workaround to handle schema with duplicate queries (#401)
* Attempt to override naming convention for ProjectMilestoneQuery

* Attempt to override naming convention for ProjectMilestoneQuery

* Deduplicate query args type names

We want to replace the `ProjectMilestone(s)?` query with
`projectMilestone(s)?`. To do this, we'll have a period where both
queries are in our API schema, which generates duplicate type names and
fails to build.

As a hack, to work around this, we can simply append a counter to those
known-duplicate type names when we encounter them for the (N > 0)-th
time, and so let the SDK build. The types are identical otherwise so
there's no functional difference, just a bit of noise in the generated
output.

* docs(changeset): override naming convention to allow support of duplicate queries

---------

Co-authored-by: Sean McGivern <sean@linear.app>